Adds the ability to launch a local container registry for devel
In order to build manifests for cross-platform images, you need to
publish architecture-specific tags for an image to a docker
registry. Then you can create a third image as a manifest that
delegates each architecture to a specific tag. When someone uses the
united manifest as the `FROM` image and tag in a Docker/Contianerfile,
the architecture used on the native device requesting the pull will
look at the manifest and download the tag associated with that architecture.